refactored the BaseResolver and finalized the GithubResolver. still needs tested
