+++ DidAuthDemo.Maui/Resolvers/BaseResolver.cs
@@ -2,9 +2,14 @@
 using CardanoSharp.Wallet.Models.Keys;
 using DidAuthDemo.Maui.Data;
 using DidAuthDemo.Maui.Models;
+using System.Net.Http.Json;
+using System;
 using System.Text.Json;
+using SimpleBase;
+using System.Text;
 
 namespace DidAuthDemo.Maui.Resolvers;
+public record DownloadDidDocumentResponse(bool IsSuccessful, DidDocument DidDocument);
 
 public interface IBaseResolver
 {
@@ -37,4 +42,38 @@ public abstract class BaseResolver : IBaseResolver
 
         return key;
     }
+
+    protected async Task<DownloadDidDocumentResponse> DownloadDidDocument(UriBuilder uriBuilder)
+    {
+        using var client = new HttpClient();
+
+        HttpResponseMessage res = await client.GetAsync(uriBuilder.Uri);
+        DidDocument didDocument;
+        if (res.IsSuccessStatusCode)
+        {
+            didDocument = await res.Content.ReadFromJsonAsync<DidDocument>();
+        }
+        else
+        {
+            string msg = await res.Content.ReadAsStringAsync();
+            Console.WriteLine(msg);
+            return new DownloadDidDocumentResponse(false, null);
+        }
+
+        if (didDocument is null) return new DownloadDidDocumentResponse(false, null);
+
+        return new DownloadDidDocumentResponse(true, didDocument);
+    }
+
+    protected bool VerifyDidDocument(DidDocument didDocument, Did did, Key key)
+    {
+        var privateKey = JsonSerializer.Deserialize<PrivateKey>(key.PrivateKey);
+        var publicKeyObj = didDocument.PublicKeys
+            .FirstOrDefault(x => x.Id == $"{did.Identifier}#key-1");
+        var publicKey = new PublicKey(Base58.Bitcoin.Decode(publicKeyObj.PublicKeyBase58), null);
+        var message = Encoding.UTF8.GetBytes("message");
+        var signature = privateKey.Sign(message);
+        return publicKey.Verify(message, signature);
+    }
 }
+
